New lemmatizer
Now based on Spacy.

Uses the transformer model, which takes some time to load, adding
considerable startup cost. For this reason I've made two unhappy
hacks in the previous "small independent tools running in parallel"
approach:
1) Built in an optional entity stripper within the lemmatizer,
rather than doubling the size of the problem by having two tools
loading the model per file. Required a makefile hack.
2) Hacked the makefile to build lemmatized target as a group.

An alternative to this would be to not use the transformer model,
but I shall stick with it for now.